+++ pUIng/src/app/components/tabs-log/new-log-section/new-log-section.component.ts
@@ -16,4 +16,6 @@ export class NewLogSectionComponent {
       ? target.value.trim()
       : target.value.trimStart()
   }
+
+  // TODO : ng new task logic
 }
