+++ pUIng/src/app/components/tabs-log/tabs-log.component.html
@@ -14,5 +14,6 @@
         </div>
         <pui-log-day *ngFor="let day of uiStateService.getLogTree() | keyvalue" [logDay]="day.key" [tree]="day.value"></pui-log-day>
         <pui-new-log-section [logsExist]="uiStateService.logsExist()"></pui-new-log-section>
+        <div #scrollAnchor></div>
     </main>
 </div>
