+++ pUIng/src/main.server.ts
@@ -1,5 +1,5 @@
 import { bootstrapApplication } from '@angular/platform-browser';
-import { AppComponent } from './app/app-root/app.component';
+import { AppComponent } from './app/components/app-root/app.component';
 import { config } from './app/app.config.server';
 
 const bootstrap = () => bootstrapApplication(AppComponent, config);
